backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
asyncpg==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
